[skip ci] Moves stokes into model
